+++ packages/terra-form-textarea/tests/wdio/form-textarea-spec.js
@@ -246,6 +246,10 @@ describe('Form-Textarea', () => {
     describe('Resizes when given several rows of input', () => {
       before(() => {
         browser.url('/#/raw/tests/terra-form-textarea/form-textarea/auto-resizable-textarea');
+        browser.execute(() => {
+          // Removes the blinking cursor to prevent screenshot mismatches.
+          document.querySelector('textarea').style.caretColor = 'transparent';
+        });
       });
 
       Terra.should.matchScreenshot('0', { viewports });
@@ -260,6 +264,10 @@ describe('Form-Textarea', () => {
     describe('Does not resize textareas that are non-resizable', () => {
       before(() => {
         browser.url('/#/raw/tests/terra-form-textarea/form-textarea/medium-textarea');
+        browser.execute(() => {
+          // Removes the blinking cursor to prevent screenshot mismatches.
+          document.querySelector('textarea').style.caretColor = 'transparent';
+        });
       });
 
       Terra.should.matchScreenshot('0', { viewports });
@@ -274,6 +282,10 @@ describe('Form-Textarea', () => {
     describe('Blocks autoresizing for when ontouch exists and browser screen is not desktop size', () => {
       before(() => {
         browser.url('/#/raw/tests/terra-form-textarea/form-textarea/auto-resizable-textarea-mobile-on-touch-start');
+        browser.execute(() => {
+          // Removes the blinking cursor to prevent screenshot mismatches.
+          document.querySelector('textarea').style.caretColor = 'transparent';
+        });
       });
 
       Terra.should.matchScreenshot('0', { viewports });
@@ -288,6 +300,10 @@ describe('Form-Textarea', () => {
     describe('Blocks autoresizing for when DocumentTouch exists and browser screen is not desktop size', () => {
       before(() => {
         browser.url('/#/raw/tests/terra-form-textarea/form-textarea/auto-resizable-textarea-mobile-document-touch');
+        browser.execute(() => {
+          // Removes the blinking cursor to prevent screenshot mismatches.
+          document.querySelector('textarea').style.caretColor = 'transparent';
+        });
       });
 
       Terra.should.matchScreenshot('0', { viewports });
@@ -302,6 +318,10 @@ describe('Form-Textarea', () => {
     describe('Blocks autoresizing for when maxTouchPoints exists and browser screen is not desktop size', () => {
       before(() => {
         browser.url('/#/raw/tests/terra-form-textarea/form-textarea/auto-resizable-textarea-mobile-max-touch-points');
+        browser.execute(() => {
+          // Removes the blinking cursor to prevent screenshot mismatches.
+          document.querySelector('textarea').style.caretColor = 'transparent';
+        });
       });
 
       Terra.should.matchScreenshot('0', { viewports });
@@ -316,6 +336,10 @@ describe('Form-Textarea', () => {
     describe('Blocks autoresizing for when msMaxTouchPoints exists and browser screen is not desktop size', () => {
       before(() => {
         browser.url('/#/raw/tests/terra-form-textarea/form-textarea/auto-resizable-textarea-mobile-ms-max-touch-points');
+        browser.execute(() => {
+          // Removes the blinking cursor to prevent screenshot mismatches.
+          document.querySelector('textarea').style.caretColor = 'transparent';
+        });
       });
 
       Terra.should.matchScreenshot('0', { viewports });