+++ packages/terra-form-select/CHANGELOG.md
@@ -3,7 +3,7 @@ ChangeLog
 
 Unreleased
 ----------
-### Fixed
+### Added
 * Added CSS style to display the select box-shadow over the toggle border
 * Added CSS custom property to enable theming of disabled state arrow icon