+++ packages/terra-form-select/CHANGELOG.md
@@ -6,6 +6,7 @@ Unreleased
 ### Added
 * Added CSS style to display the select box-shadow over the toggle border
 * Added CSS custom property to enable theming of disabled state arrow icon
+* Null check the active state before performing findByValue
 
 4.7.0 - (June 28, 2018)
 ------------------