+++ packages/terra-form-select/CHANGELOG.md
@@ -3,6 +3,17 @@ ChangeLog
 
 Unreleased
 ----------
+### Fixed
+* Added CSS style to display the select box-shadow over the toggle border
+* Null check the active state before performing findByValue
+
+4.7.0 - (June 28, 2018)
+------------------
+### Changed
+* Minor dependency version bump
+
+### Added
+* Added change log to the documentation on terra-dev-site
 
 4.6.0 - (June 22, 2018)
 ------------------